docker/requirements.txt
web3==6.*
oasis-sapphire-py
python-dotenv
pycryptodome
